Refactor: lrmd: separate alert API handler into its own file
lrmd.c is too big
