+++ lrmd/Makefile.am
@@ -39,7 +39,7 @@ lrmd_LDFLAGS		= $(LDFLAGS_HARDENED_EXE)
 lrmd_LDADD		= $(top_builddir)/lib/common/libcrmcommon.la	\
 			$(top_builddir)/lib/services/libcrmservice.la	\
 			$(top_builddir)/lib/fencing/libstonithd.la ${COMPAT_LIBS}
-lrmd_SOURCES		= main.c lrmd.c
+lrmd_SOURCES		= main.c lrmd.c lrmd_alert_api.c
 
 pacemaker_remoted_CPPFLAGS	= -DSUPPORT_REMOTE $(AM_CPPFLAGS)
 
@@ -48,7 +48,7 @@ pacemaker_remoted_LDFLAGS	= $(LDFLAGS_HARDENED_EXE)
 
 pacemaker_remoted_LDADD		= $(lrmd_LDADD) \
 			$(top_builddir)/lib/lrmd/liblrmd.la
-pacemaker_remoted_SOURCES	= main.c lrmd.c tls_backend.c ipc_proxy.c
+pacemaker_remoted_SOURCES	= $(lrmd_SOURCES) tls_backend.c ipc_proxy.c
 
 lrmd_internal_ctl_LDADD		= $(top_builddir)/lib/common/libcrmcommon.la	\
 				$(top_builddir)/lib/lrmd/liblrmd.la		\
