+++ lrmd/lrmd_private.h
@@ -111,5 +111,5 @@ crm_client_t *ipc_proxy_get_provider(void);
 int ipc_proxy_shutdown_req(crm_client_t *ipc_proxy);
 #endif
 
+int process_lrmd_alert_exec(crm_client_t *client, uint32_t id, xmlNode *request);
 #endif
-
